src/requirements.txt
discord
mysql-connector-python
watchdog
alembic
python-dotenv
pymysql
cryptography